refactor(repo): extract the @intake/ai-prompts package (phase 4a)
Phase 4a of the Turborepo migration: pull the app's SDK-free AI
prompt/tool artifacts into a new pure package, @intake/ai-prompts
(deps: @intake/types + zod only), consumed as raw TS via
transpilePackages. The Anthropic SDK client, API-key vault, usage
tracker, and route-level zod request/response validation stay in
apps/web.

What moved:
- CLAUDE_MODELS + WEB_SEARCH_TOOL registry -> @intake/ai-prompts/models;
  claude-client.ts re-exports them so all 10 ai routes + bug-report +
  their tests resolve unchanged.
- analytics-insights.ts -> @intake/ai-prompts/analytics-insights (zod
  schemas + INSIGHT_TOOL + INSIGHTS_SYSTEM_PROMPT + buildInsightsPrompt),
  DOMAINS/Domain import repointed to @intake/types/analytics, and the
  deep-mode DEEP_SYSTEM_PROMPT folded in beside its ${INSIGHTS_SYSTEM_PROMPT}
  base. Re-export shim at @/lib/analytics-insights keeps all importers
  unchanged. (Clears the DEFERRED.md Phase 3 item.)
- Every api/ai/** route's system prompt + tool-def (parse,
  medicine-search, substance-lookup, substance-enrich, interaction-check,
  nutrient-analysis, titration-warnings, voice-parse). The two schema.ts
  routes re-export their tool-def from the package so route/schema tests
  stay green.

All moved prompt/tool text is byte-identical to before (no behaviour
change). New ai-prompts-purity.test.ts asserts the package never imports
@anthropic-ai/sdk / server-only / @intake/db / next / react (incl.
dynamic import()/require()) — load-bearing because the analytics-insights
schemas are transitively reachable from the client bundle.

Verified: turbo typecheck (5 pkgs), lint (0 errors), 2026 unit tests,
33 integration (testcontainers), production build, bundle-security scan,
3-agent adversarial review (byte-fidelity / boundary / wiring) = clean.
