+++ apps/web/next.config.mjs
@@ -45,8 +45,13 @@ const nextConfig = {
   output: isCapacitorBuild ? "export" : undefined,
   reactStrictMode: true,
   // Internal @intake/* packages are consumed as raw TS source (JIT). Grows as
-  // core, ui, ai-prompts are extracted in later phases.
-  transpilePackages: ["@intake/core", "@intake/db", "@intake/types"],
+  // ui is extracted in a later phase.
+  transpilePackages: [
+    "@intake/ai-prompts",
+    "@intake/core",
+    "@intake/db",
+    "@intake/types",
+  ],
   // Pin the Turbopack root to the monorepo root so module resolution + the
   // version require above are scoped correctly (silences the inferred-root warning).
   turbopack: {
