+++ apps/web/src/app/api/ai/_shared/claude-client.ts
@@ -23,17 +23,11 @@ export async function getClaudeClientForUser(
   return { client, resolved };
 }
 
-export const CLAUDE_MODELS = {
-  fast: "claude-haiku-4-5-20251001" as const,
-  quality: "claude-sonnet-4-6" as const,
-  premium: "claude-opus-4-8" as const,
-} as const;
-
-export const WEB_SEARCH_TOOL = {
-  type: "web_search_20250305" as const,
-  name: "web_search" as const,
-  max_uses: 5,
-};
+// The pinned model registry + shared web-search tool now live in the
+// SDK-free @intake/ai-prompts package; re-exported here so every route's
+// `import { ..., CLAUDE_MODELS, WEB_SEARCH_TOOL } from ".../claude-client"`
+// resolves unchanged.
+export { CLAUDE_MODELS, WEB_SEARCH_TOOL } from "@intake/ai-prompts/models";
 
 export {
   GRAMS_PER_STANDARD_DRINK,
