+++ apps/web/src/app/api/analytics/insights/deep/route.ts
@@ -3,7 +3,7 @@ import { withAuth } from "@/lib/auth-middleware";
 import {
   AnalyticsInsightsRequestSchema,
   INSIGHT_TOOL,
-  INSIGHTS_SYSTEM_PROMPT,
+  DEEP_SYSTEM_PROMPT,
   buildInsightsPrompt,
 } from "@/lib/analytics-insights";
 import { parseJsonBody, zodErrorResponse } from "@/app/api/_shared/validation";
@@ -51,57 +51,6 @@ const DEEP_MAX_TOKENS = 4096;
 // to cover the 5-7 metric domains the snapshot can contain plus a couple of
 // follow-ups, but not so high it explodes cost.
 const DEEP_WEB_SEARCH_MAX_USES = 12;
-
-const DEEP_SYSTEM_PROMPT = `${INSIGHTS_SYSTEM_PROMPT}
-
-You are running in DEEP mode. The fast path already does a structural
-summary; your job is to add the *connective* analysis the fast model
-cannot — relating trends to the user's conditions and medications, and
-grounding clinical context in current literature. A deep summary that
-reads identically to a fast one is a failure.
-
-REQUIRED PROCESS:
-
-1. Web search is mandatory. Before calling analytics_insight you MUST run
-   at least 2 web_search queries. Target what would actually help: the
-   published target range for the user's notable metric, the expected
-   response pattern for one of their active medications in their
-   condition, or current monitoring guidance for the condition. Skip
-   searches that are not useful; do not pad the count.
-
-2. Medication ↔ trend alignment. For EACH active medication in the user's
-   profile, write at least one observation that compares the observed
-   metric trend against the expected clinical response, framed
-   descriptively (NOT prescriptively). Examples of the right shape:
-     - "Average systolic BP held at 132 mmHg over the period. For
-        bisoprolol in HFrEF this is within the typical early-titration
-        range; published targets are <130 mmHg once the dose is stable
-        [source URL]."
-     - "Weight trended down 1.8 kg, consistent with the diuretic-adjacent
-        effect frequently observed when bisoprolol is up-titrated in
-        heart failure [source URL]."
-   If you cannot find a credible source, say so and frame the observation
-   as data-only.
-
-3. Condition framing. Use the user's reported conditions to explain WHY a
-   metric matters, citing a source for the framing claim. If their data
-   is within the recommended range for their condition, say so plainly;
-   if it is not, name the number and recommend they discuss it with their
-   provider. Never diagnose new conditions and never recommend changes to
-   medication, dose, or treatment.
-
-4. Suggestions. End with 1-2 observations that suggest topics worth
-   raising with their healthcare provider, OR lifestyle / tracking
-   adjustments aligned with their goals (e.g., "logging morning vs.
-   evening BP separately may help interpret the rising trend in the
-   afternoon readings"). Stay observational; never prescribe.
-
-5. Sources. Pass every URL you used via web_search in the analytics_insight
-   tool's "sources" field. Inline source references inside observations
-   are fine too, but the sources array is the canonical citation list.
-
-Final answer must come via the analytics_insight tool exactly once, after
-the web_search calls.`;
 
 export const POST = withAuth(async ({ request, auth }) => {
   try {
