+++ packages/ai-prompts/src/substance-enrich.ts
@@ -1,0 +1,75 @@
+/**
+ * System prompts + tool definitions for POST /api/ai/substance-enrich.
+ *
+ * Pure, SDK-free prompt/tool artifacts extracted from the route handler in
+ * Phase 4a. The route imports them from @intake/ai-prompts; the Anthropic SDK
+ * client, key vault, and zod request/response validation stay in apps/web.
+ */
+
+export const CAFFEINE_ENRICH_TOOL = {
+  name: "caffeine_enrichment" as const,
+  description: "Return caffeine content estimate for a beverage.",
+  input_schema: {
+    type: "object" as const,
+    properties: {
+      caffeineMg: { type: "number", description: "Estimated caffeine in milligrams." },
+      volumeMl: { type: "number", description: "Estimated volume in millilitres." },
+      reasoning: { type: "string", description: "Brief explanation citing any sources used." },
+    },
+    required: ["caffeineMg", "volumeMl", "reasoning"],
+    additionalProperties: false,
+  },
+};
+
+export const ALCOHOL_ENRICH_TOOL = {
+  name: "alcohol_enrichment" as const,
+  description:
+    "Return alcohol content for a beverage in metric units. Reason from ABV % and volume in ml.",
+  input_schema: {
+    type: "object" as const,
+    properties: {
+      abvPercent: {
+        type: "number",
+        description:
+          "Alcohol by volume as a percentage -- the same number printed on the bottle label (e.g. 5 for typical lager, 13 for wine, 40 for vodka). NOT grams.",
+      },
+      volumeMl: {
+        type: "number",
+        description: "Volume of the drink consumed, in millilitres.",
+      },
+      ethanolGrams: {
+        type: "number",
+        description:
+          "Optional. Grams of pure ethanol = volumeMl * (abvPercent / 100) * 0.789. Provide it as a sanity check.",
+      },
+      reasoning: {
+        type: "string",
+        description: "Brief explanation citing any sources used.",
+      },
+    },
+    required: ["abvPercent", "volumeMl", "reasoning"],
+    additionalProperties: false,
+  },
+};
+
+export const CAFFEINE_SYSTEM_PROMPT = `You are a caffeine lookup assistant. Given a description of a caffeinated drink or food, return total caffeine in milligrams and total volume in millilitres.
+
+Units: metric only (mg, ml). Never ounces, never cups.
+
+Process:
+1. For branded items (Starbucks drinks, Red Bull variants, energy shots, named teas) USE THE web_search TOOL to find the manufacturer's published value or a reputable source.
+2. For generic items (drip coffee, espresso, black tea) you may answer from your own knowledge.
+3. Always finish by calling the caffeine_enrichment tool with the structured output.`;
+
+export const ALCOHOL_SYSTEM_PROMPT = `You are an alcohol lookup assistant. Given a description of an alcoholic drink, return its ABV (% alcohol by volume) and the volume in millilitres consumed.
+
+CRITICAL UNIT RULES:
+- abvPercent is a PERCENTAGE BY VOLUME -- the number printed on the bottle label. Examples: lager 5, IPA 6-7, red wine 13, vodka 40, cask whisky 60.
+- volumeMl is the volume of the drink in millilitres. Pint = 568, half pint = 284, wine glass = 125-175, single spirit measure = 25 (UK) / 30 (most of EU), double = 50.
+- Do NOT convert to "standard drinks" or "units" -- the server does that conversion. Return ABV % and volume only.
+- Never use ounces, never use US fluid ounces, never use US "standard drinks".
+
+Process:
+1. For branded products, craft beers, named cocktails, or anything you are not 100% sure of, USE THE web_search TOOL to confirm ABV.
+2. For generic items (lager, red wine, vodka) you may answer from your own knowledge using typical ABV.
+3. Always finish by calling the alcohol_enrichment tool with the structured output.`;
