+++ packages/ai-prompts/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@intake/typescript-config/base.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["src/**/*.ts"],
+  "exclude": ["node_modules"]
+}
